+++ .setup/r-dependencies/processLM/processLM-process-lm-dot.R
@@ -60,24 +60,12 @@
     sigmacap_consistent
   )
   pinv_of_dcap <- .PInvDmat(.DMat(k))
-  if (p > 1) {
-    dif_idx <- utils::combn(seq_len(p), 2)
-    p_dif <- dim(dif_idx)[2]
-    dif_betastar <- rep(x = 0.0, times = p_dif)
-    dif_beta <- rep(x = 0.0, times = p_dif)
-    dif_names <- rep(x = 0.0, times = p_dif)
-    for (i in seq_len(p_dif)) {
-      dif_betastar[i] <- betastar[dif_idx[1, i]] - betastar[dif_idx[2, i]]
-      dif_beta[i] <- beta[dif_idx[1, i]] - beta[dif_idx[2, i]]
-      dif_names[i] <- paste0(xnames[dif_idx[1, i]], "-", xnames[dif_idx[2, i]])
-    }
-    names(dif_betastar) <- dif_names
-    names(dif_beta) <- dif_names
-  } else {
-    dif_betastar <- NULL
-    dif_beta <- NULL
-    dif_idx <- NULL
-  }
+  dif <- .Dif(
+    beta = beta,
+    betastar = betastar,
+    p = p,
+    xnames = xnames
+  )
   rsq <- c(
     rsq = summary_lm$r.squared,
     adj = summary_lm$adj.r.squared
@@ -106,9 +94,9 @@
     beta = beta,
     sigmasq = sigmasq,
     theta = unname(theta),
-    dif_betastar = dif_betastar,
-    dif_beta = dif_beta,
-    dif_idx = dif_idx,
+    dif_betastar = dif$dif_betastar,
+    dif_beta = dif$dif_beta,
+    dif_idx = dif$dif_idx,
     rsq = rsq
   )
 }